+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.8.0 [unreleased]
+* **[breaking]** Removed support for `HTTPVaultConnectorFactory#withSslContext()` in favor of `#withTrustedCA()` due to 
+refactoring of the internal SSL handling.
+* [improvement] `VaultConnector` extends `java.io.Serializable`
+
 ## 0.7.1 [2018-03-17]
 * [improvement] Added automatic module name for JPMS compatibility
 * [dependencies] Minor dependency updates